Fixed warnings.

+++ Source/Player.cpp
@@ -259,7 +259,7 @@ void Player::configureChannels()
 {
 	if (m_channelMappingWindow.get() == nullptr) {
 		m_channelMappingWindow.set(new ChannelMappingWindow(m_outputChannelNames, m_tracksContainer.createMapping(), [&](int source, int target) {
-			for (int i = 0; i < m_tracksContainer.size(); ++i) {
+			for (size_t i = 0; i < m_tracksContainer.size(); ++i) {
 				if (source - m_tracksContainer[i].getNumChannels() < 0) {
 					m_tracksContainer[i].setOutputChannelMapping(source, target);
 					break;
